delegate cannot be static

+++ test/test_hashtable.cpp
@@ -359,7 +359,7 @@ const char* test_hashtable()
         test_count
     };
 
-    for (int i = 0; i < KJSD_LENGTH(f); i++)
+    for (size_t i = 0; i < KJSD_LENGTH(f); i++)
     {
         setUp();
         KJSD_CUNIT_RUN(f[i]);
