+++ test/test_state_machine.cpp
@@ -177,7 +177,7 @@ const char* test_state_machine()
         test_passive
     };
 
-    for (int i = 0; i < KJSD_LENGTH(f); i++)
+    for (size_t i = 0; i < KJSD_LENGTH(f); i++)
     {
         setUp();
         KJSD_CUNIT_RUN(f[i]);
